Create pyadept package
This commit refactors the sources and creates an installable python package.

Notable changes include:
- Move tests to a dedicated folder
- Reformat all files using black
- Manage dependencies using poetry
- Set-up continuous integration using github actions
